localization proj trial 1

+++ Lesson_6_Utilizing_Scan_Matching/Exercise-ICP-Alignment/helper.cpp
@@ -50,7 +50,7 @@ double getDistance(Point p1, Point p2){
 double minDistance(Point p1, vector<Point> points){
 	if(points.size() > 0){
 		double dist = getDistance(p1, points[0]);
-		for(uint index = 1; index < points.size(); index++){
+		for(size_t index = 1; index < points.size(); index++){
 			double newDist = getDistance(p1, points[index]);
 			if( newDist < dist)
 				dist = newDist;
@@ -91,7 +91,7 @@ void renderRay(pcl::visualization::PCLVisualizer::Ptr& viewer, Point p1, Point p
 void renderPath(pcl::visualization::PCLVisualizer::Ptr& viewer, const PointCloudT::Ptr& cloud, std::string name, Color color){
 
 	int previous = 0;
-	for(uint index = previous+1; index < cloud->points.size(); index++){
+	for(size_t index = previous+1; index < cloud->points.size(); index++){
 		renderRay(viewer, Point(cloud->points[previous].x, cloud->points[previous].y, 0), Point(cloud->points[index].x, cloud->points[index].y, 0), name+to_string(previous), color);
 		previous++;
 	}
